docs: fix chromatic layout issues
