+++ src/components/drawer/bl-drawer.stories.mdx
@@ -1,17 +1,20 @@
 import { html } from 'lit';
 import { ifDefined } from 'lit/directives/if-defined.js';
 import { Meta, Canvas, ArgsTable, Story } from '@storybook/addon-docs';
+import { fullscreenLayout, withNoAnimation } from '../../utilities/chromatic-decorators';
 
 <Meta
   title="Components/Drawer"
   component="bl-drawer"
   parameters={{
-    layout: 'fullscreen',
-    noAnimationOnChromaticLayout: true,
     chromatic: {
       viewports: [768, 320],
     },
   }}
+  decorators={[
+    fullscreenLayout,
+    withNoAnimation,
+  ]}
   argTypes={{
     open: {
       control: "boolean",
